Support setting a "legal value range" in each metric config
If your metric is e.g. "list-database-users | wc -l" and the command
happens to run with VPN disconnected, you probably don't want to
record a "0" value just because stdout of "list-database-users"
was empty (maybe its stderr contained a 'Cannot connect to db' error).

After this commit, it is possible to add a "require" key in the
metric config, inside which above/below constraints on the value
can be specified. In the future, support might be added for
constraints on the exitcode and/or time taken of the metric
command in parallel to the constraints on the value it outputs.
It's possible to add just an "above" constraint or just a "below"
constraint, or both.

Example metric config:

{
  "cmd": "some-command-here",
  "title": "some title here",
  "require": {
    "value": {
      "above": 0,
      "below": 100
    }
  }
}